add resources to sidebar
