+++ app/routes/_layout-app/_mini-projetos/mini-projetos_.$slug_/_tabs/_overview/components/join-challenge-section.tsx
